AppTimTro/main/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.8
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==2.2.0
django-oauth-toolkit==2.4.0
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.6
myapp==0.1.dev0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
pycparser==2.22
PyMySQL==1.1.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
